LayoutTests/editing/pasteboard/entries-api/datatransfer-items-drop-getAsEntry-expected.txt

Basic test coverage for dataTransferItem.webkitGetAsEntry()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS dataTransfer.items.length is 2
PASS firstDataTransferItem.kind is "file"
PASS firstDataTransferItem.type is "text/plain"
PASS firstFile.name is "test.txt"
PASS firstFile.size is 5
PASS firstFile.type is "text/plain"
PASS firstEntry.isDirectory is false
PASS firstEntry.isFile is true
PASS firstEntry.name is "test.txt"
PASS firstEntry.fullPath is "/test.txt"
PASS secondDataTransferItem.kind is "file"
PASS secondDataTransferItem.type is ""
PASS secondFile.name is "testFiles"
PASS secondFile.size is one of `272`, and `256`
PASS secondFile.type is ""
PASS secondEntry.isDirectory is true
PASS secondEntry.isFile is false
PASS secondEntry.name is "testFiles"
PASS secondEntry.fullPath is "/testFiles"
PASS fileSystemRoot.isDirectory is true
PASS fileSystemRoot.isFile is false
PASS fileSystemRoot.name is ""
PASS fileSystemRoot.fullPath is "/"
PASS successfullyParsed is true

TEST COMPLETE
